Fix rspec dependencies
We are using a super old version of puppet-tempest, which depends on
openstack/puppet-openstack_spec_helper that has moved on to depend on
rspec 2.3.0

Update rspec version in this Gemfile for now, so that tests pass and we
can merge a few dependent changes.

TODO: We should pull in latest upstream puppet-tempest and use that
instead of maintaining this old version of puppet-tempest.

Change-Id: Icce5534eff99816c8189141a598050e24b2b4718